メディア作成フォーム： - 編集ページにも使えるようにする - コンポーネント化
